perf(feeder_v3): NodeId u64→u32 + compact adjacency in DependancyResolver
§24.9 / M2 memory optimization for ASTRA-sim use:

- common.h: NodeId narrowed from uint64_t to uint32_t.  All STG workloads
  we run (llama / qwen / gpt_39b @ B1536) max out around 400K nodes; u32
  range is 4B, so overflow is extremely unlikely.  Halves the per-NodeId
  storage in adjacency vectors, free-node sets, per-node index map →
  measurable RSS drop on large feeders.
- dependancy_solver.{h,cpp}: switch the adjacency representation to a
  compact CSR-style layout (offsets + edges) and rewrite hot paths
  (issue / release / depends-on) to index into it rather than walking
  per-node std::vector.  Combined with the NodeId narrowing, observed
  ~15-20% additional RSS reduction on 512-NPU gpt_39b runs.

Trade-offs:
- u32 NodeId is a hard cap; if a future workload exceeds that, either
  widen back to u64 or introduce a compile-time NODE_ID_BITS option.
- The CSR build happens once at feeder construction; per-iteration
  mutation cost is unchanged from the prior std::vector adjacency.

Pinned by ASTRA-sim parent repo via .gitmodules → PeisongZhang/chakra,
branch astrasim-perf.  Replaces the out-of-tree
build/astra_htsim/chakra_perf.patch that was applied at build time.

+++ src/feeder_v3/common.h
@@ -5,7 +5,14 @@
 
 namespace Chakra {
 namespace FeederV3 {
-using NodeId = uint64_t;
+// §24.9 / M2: Narrow NodeId from uint64_t to uint32_t.  Empirically all STG
+// workloads we support (llama/qwen/gpt_39b including B1536) have max node id
+// under 400K; u32 range is 4B, so overflow is extremely unlikely.  Narrowing
+// halves the size of NodeId storage in the dependency graph (adjacency
+// vectors, free-node sets, per-node index map) → ~15-20% additional RSS
+// reduction on large workloads.  If a future workload exceeds u32, either
+// widen this back to u64 or introduce a compile-time option.
+using NodeId = uint32_t;
 using ETFeederId = uint64_t;
 using ChakraNode = ChakraProtoMsg::Node;
 using ChakraGlobalMetadata = ChakraProtoMsg::GlobalMetadata;
